+++ src/mapping/impl/net/minecraft/util/AxisAlignedBB.h
@@ -7,18 +7,18 @@
 
 
 #include "../../../../AbstractClass.h"
-#include "MovingObjectPosition.h"
+#include <net/minecraft/util/MovingObjectPosition.h>
 
 class AxisAlignedBB : public AbstractClass {
 public:
     AxisAlignedBB(Phantom *phantom, jobject aabb);
 
-    jobject expand(jfloat x, jfloat y, jfloat z);
+    jobject expand(jdouble x, jdouble y, jdouble z);
     jobject addCoord(jdouble x, jdouble y, jdouble z);
     jobject calculateIntercept(jobject vec1, jobject vec2);
     jboolean isVecInside(jobject vec);
 
-    AxisAlignedBB *getExpandContainer(jfloat x, jfloat y, jfloat z);
+    AxisAlignedBB *getExpandContainer(jdouble x, jdouble y, jdouble z);
     AxisAlignedBB *getAddCoordContainer(jdouble x, jdouble y, jdouble z);
     MovingObjectPosition *getCalculateInterceptContainer(jobject vec1, jobject vec2);
 private:
